コマンド実行関係を cmd_exec_command.c と cmd_exec_commands.c に移動
